refactor(api): split notifyAllUsersInSubTools from user service
